+++ src/main/java/graphql/schema/GraphQLArgument.java
@@ -148,7 +148,7 @@ public class GraphQLArgument implements GraphQLDirectiveContainer {
     }
 
     @Override
-    public TraversalControl accept(TraverserContext<GraphQLType> context, TypeVisitor visitor) {
+    public TraversalControl accept(TraverserContext<GraphQLType> context, GraphqlTypeVisitor visitor) {
         return visitor.visitGraphQLArgument(this, context);
     }
