+++ src/main/java/graphql/schema/GraphQLEnumValueDefinition.java
@@ -108,7 +108,7 @@ public class GraphQLEnumValueDefinition implements GraphQLDirectiveContainer {
     }
 
     @Override
-    public TraversalControl accept(TraverserContext<GraphQLType> context, TypeVisitor visitor) {
+    public TraversalControl accept(TraverserContext<GraphQLType> context, GraphqlTypeVisitor visitor) {
         return visitor.visitGraphQLEnumValueDefinition(this, context);
     }
