+++ src/main/java/graphql/schema/GraphQLFieldDefinition.java
@@ -152,7 +152,7 @@ public class GraphQLFieldDefinition implements GraphQLDirectiveContainer {
     }
 
     @Override
-    public TraversalControl accept(TraverserContext<GraphQLType> context, TypeVisitor visitor) {
+    public TraversalControl accept(TraverserContext<GraphQLType> context, GraphqlTypeVisitor visitor) {
         return visitor.visitGraphQLFieldDefinition(this, context);
     }
