+++ src/main/java/graphql/schema/GraphQLInputObjectField.java
@@ -111,7 +111,7 @@ public class GraphQLInputObjectField implements GraphQLDirectiveContainer {
     }
 
     @Override
-    public TraversalControl accept(TraverserContext<GraphQLType> context, TypeVisitor visitor) {
+    public TraversalControl accept(TraverserContext<GraphQLType> context, GraphqlTypeVisitor visitor) {
         return visitor.visitGraphQLInputObjectField(this, context);
     }
