+++ src/main/java/graphql/schema/GraphQLScalarType.java
@@ -108,7 +108,7 @@ public class GraphQLScalarType implements GraphQLType, GraphQLInputType, GraphQL
     }
 
     @Override
-    public TraversalControl accept(TraverserContext<GraphQLType> context, TypeVisitor visitor) {
+    public TraversalControl accept(TraverserContext<GraphQLType> context, GraphqlTypeVisitor visitor) {
         return visitor.visitGraphQLScalarType(this, context);
     }
