+++ src/main/java/graphql/schema/GraphQLTypeReference.java
@@ -41,7 +41,7 @@ public class GraphQLTypeReference implements GraphQLType, GraphQLOutputType, Gra
     }
 
     @Override
-    public TraversalControl accept(TraverserContext<GraphQLType> context, TypeVisitor visitor) {
+    public TraversalControl accept(TraverserContext<GraphQLType> context, GraphqlTypeVisitor visitor) {
         return visitor.visitGraphQLTypeReference(this, context);
     }
 }
