+++ src/main/java/graphql/schema/SchemaUtil.java
@@ -2,11 +2,13 @@ package graphql.schema;
 
 
 import graphql.Assert;
-import graphql.AssertException;
 import graphql.Internal;
 import graphql.introspection.Introspection;
 import graphql.util.TraversalControl;
 import static graphql.util.TraversalControl.QUIT;
+import static graphql.schema.GraphqlTypeVisitors.LeafVisitor;
+import static graphql.schema.GraphqlTypeVisitors.CollectingVisitor;
+import static graphql.schema.GraphqlTypeVisitors.TypeResolvingVisitor;
 import graphql.util.TraverserContext;
 
 import java.util.ArrayList;
@@ -15,7 +17,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collectors;
 
 import static graphql.schema.GraphQLTypeUtil.isList;
 import static graphql.schema.GraphQLTypeUtil.isNonNull;
@@ -25,11 +26,17 @@ import static java.lang.String.format;
 @Internal
 public class SchemaUtil {
 
+    private static final TypeTraverser SHALLOW_TRAVERSER = new TypeTraverser((node) -> Collections.emptyList());
+
+    private static final TypeTraverser TRAVERSER = new TypeTraverser();
+
+
     public Boolean isLeafType(GraphQLType graphQLType) {
         return (Boolean) SHALLOW_TRAVERSER.depthFirst(new LeafVisitor(), graphQLType).getResult();
     }
 
     public boolean isInputType(GraphQLType graphQLType) {
+
         return (Boolean) SHALLOW_TRAVERSER.depthFirst(new LeafVisitor() {
             @Override
             public TraversalControl visitGraphQLInputObjectType(GraphQLInputObjectType node, TraverserContext<GraphQLType> context) {
@@ -52,29 +59,7 @@ public class SchemaUtil {
 
     }
 
-    /*
-        From http://facebook.github.io/graphql/#sec-Type-System
 
-           All types within a GraphQL schema must have unique names. No two provided types may have the same name.
-           No provided type may have a name which conflicts with any built in types (including Scalar and Introspection types).
-
-        Enforcing this helps avoid problems later down the track fo example https://github.com/graphql-java/graphql-java/issues/373
-     */
-    private void assertTypeUniqueness(GraphQLType type, Map<String, GraphQLType> result) {
-        GraphQLType existingType = result.get(type.getName());
-        // do we have an existing definition
-        if (existingType != null) {
-            // type references are ok
-            if (!(existingType instanceof GraphQLTypeReference || type instanceof GraphQLTypeReference))
-                // object comparison here is deliberate
-                if (existingType != type) {
-                    throw new AssertException(format("All types within a GraphQL schema must have unique names. No two provided types may have the same name.\n" +
-                                    "No provided type may have a name which conflicts with any built in types (including Scalar and Introspection types).\n" +
-                                    "You have redefined the type '%s' from being a '%s' to a '%s'",
-                            type.getName(), existingType.getClass().getSimpleName(), type.getClass().getSimpleName()));
-                }
-        }
-    }
 
     Map<String, GraphQLType> allTypes(final GraphQLSchema schema,final Set<GraphQLType> additionalTypes) {
         List<GraphQLType> roots = new ArrayList<GraphQLType>() {{
@@ -158,39 +143,8 @@ public class SchemaUtil {
     }
 
     void replaceTypeReferences(GraphQLSchema schema) {
-
         final Map<String, GraphQLType> typeMap = schema.getTypeMap();
-
-        TRAVERSER.depthFirst(new TypeVisitorStub() {
-
-            @Override
-            public TraversalControl visitGraphQLObjectType(GraphQLObjectType node, TraverserContext<GraphQLType> context) {
-
-               node.replaceInterfaces(node.getInterfaces().stream()
-                        .map(type -> (GraphQLOutputType)typeMap.get(type.getName()))
-                        .collect(Collectors.toList()));
-                return super.visitGraphQLObjectType(node, context);
-            }
-
-            @Override
-            public TraversalControl visitGraphQLUnionType(GraphQLUnionType node, TraverserContext<GraphQLType> context) {
-
-                node.replaceTypes(node.getTypes().stream()
-                        .map(type -> (GraphQLOutputType)typeMap.get(type.getName()))
-                        .collect(Collectors.toList()));
-                return super.visitGraphQLUnionType(node, context);
-            }
-
-            @Override
-            public TraversalControl visitGraphQLTypeReference(GraphQLTypeReference node, TraverserContext<GraphQLType> context) {
-
-                final GraphQLType resolvedType = typeMap.get(node.getName());
-                Assert.assertTrue(resolvedType != null, "type %s not found in schema", node.getName());
-                context.getParentContext().thisNode().accept(context, new RefResolvingVisitor(resolvedType) );
-                return super.visitGraphQLTypeReference(node, context);
-            }
-        },typeMap.values());
-
+        TRAVERSER.depthFirst(new TypeResolvingVisitor(typeMap),typeMap.values());
     }
 
     @Deprecated
@@ -204,160 +158,6 @@ public class SchemaUtil {
     }
 
 
-    class RefResolvingVisitor extends TypeVisitorStub {
-
-        final private GraphQLType resolvedType;
-        public RefResolvingVisitor(GraphQLType resolvedType) {
-            this.resolvedType = resolvedType;
-        }
-
-        @Override
-        public TraversalControl visitGraphQLFieldDefinition(GraphQLFieldDefinition node, TraverserContext<GraphQLType> context) {
-            node.replaceType((GraphQLOutputType) resolvedType);
-            return super.visitGraphQLFieldDefinition(node, context);
-        }
-
-        @Override
-        public TraversalControl visitGraphQLArgument(GraphQLArgument node, TraverserContext<GraphQLType> context) {
-            node.replaceType((GraphQLInputType) resolvedType);
-            return super.visitGraphQLArgument(node, context);
-        }
-
-        @Override
-        public TraversalControl visitGraphQLInputObjectField(GraphQLInputObjectField node, TraverserContext<GraphQLType> context) {
-            node.replaceType((GraphQLInputType) resolvedType);
-            return super.visitGraphQLInputObjectField(node, context);
-        }
-
-        @Override
-        public TraversalControl visitGraphQLList(GraphQLList node, TraverserContext<GraphQLType> context) {
-            node.replaceType( resolvedType);
-            return super.visitGraphQLList(node, context);
-        }
-
-        @Override
-        public TraversalControl visitGraphQLNonNull(GraphQLNonNull node, TraverserContext<GraphQLType> context) {
-            node.replaceType( resolvedType);
-            return super.visitGraphQLNonNull(node, context);
-        }
-
-    }
-
-
-    public static class LeafVisitor extends TypeVisitorStub {
-
-        @Override
-        public TraversalControl visitGraphQLScalarType(GraphQLScalarType node, TraverserContext<GraphQLType> context) {
-            context.setResult(true);
-            return QUIT;
-        }
-
-        @Override
-        public TraversalControl visitGraphQLEnumType(GraphQLEnumType node, TraverserContext<GraphQLType> context) {
-            context.setResult(true);
-            return QUIT;
-        }
-
-        @Override
-        protected TraversalControl visitGraphQLType(GraphQLType node, TraverserContext<GraphQLType> context) {
-            context.setResult(false);
-            return QUIT;
-        }
-
-        @Override
-        public TraversalControl visitGraphQLList(GraphQLList node, TraverserContext<GraphQLType> context) {
-            return node.getWrappedType().accept(context,this);
-        }
-
-        @Override
-        public TraversalControl visitGraphQLNonNull(GraphQLNonNull node, TraverserContext<GraphQLType> context) {
-            return node.getWrappedType().accept(context,this);
-        }
-
-    }
-
-
-    class CollectingVisitor extends TypeVisitorStub {
-
-        private final Map<String,GraphQLType> result = new HashMap<>();
-
-        @Override
-        public TraversalControl visitGraphQLEnumType(GraphQLEnumType node, TraverserContext<GraphQLType> context) {
-            assertTypeUniqueness(node,result);
-            save(node.getName(),node);
-            return super.visitGraphQLEnumType(node, context);
-        }
-
-        @Override
-        public TraversalControl visitGraphQLScalarType(GraphQLScalarType node, TraverserContext<GraphQLType> context) {
-            assertTypeUniqueness(node,result);
-            save(node.getName(),node);
-            return super.visitGraphQLScalarType(node, context);
-        }
-
-        @Override
-        public TraversalControl visitGraphQLObjectType(GraphQLObjectType node, TraverserContext<GraphQLType> context) {
-            if (isTypeReference(node.getName())) {
-                assertTypeUniqueness(node, result);
-            } else {
-                save(node.getName(), node);
-            }
-            return super.visitGraphQLObjectType(node,context);
-        }
-
-        @Override
-        public TraversalControl visitGraphQLInputObjectType(GraphQLInputObjectType node, TraverserContext<GraphQLType> context) {
-            if (isTypeReference(node.getName())) {
-                assertTypeUniqueness(node, result);
-            } else {
-                save(node.getName(), node);
-            }
-            return super.visitGraphQLInputObjectType(node, context);
-        }
-
-        @Override
-        public TraversalControl visitGraphQLInterfaceType(GraphQLInterfaceType node, TraverserContext<GraphQLType> context) {
-            if (isTypeReference(node.getName())) {
-                assertTypeUniqueness(node, result);
-            } else {
-                save(node.getName(), node);
-            }
-
-            return super.visitGraphQLInterfaceType(node, context);
-        }
-
-        @Override
-        public TraversalControl visitGraphQLUnionType(GraphQLUnionType node, TraverserContext<GraphQLType> context) {
-            assertTypeUniqueness(node,result);
-            save(node.getName(),node);
-            return super.visitGraphQLUnionType(node, context);
-        }
-
-        @Override
-        public TraversalControl visitGraphQLFieldDefinition(GraphQLFieldDefinition node, TraverserContext<GraphQLType> context) {
-
-            return super.visitGraphQLFieldDefinition(node, context);
-        }
-
-        // TODO: eh? Isn't it similar to assertTypeUniqueness?
-        private boolean isTypeReference(String name) {
-            return result.containsKey(name) && !(result.get(name) instanceof GraphQLTypeReference);
-        }
-
-        private void save(String name, GraphQLType type) {
-            result.put(name,type);
-        }
-
-        public Map<String, GraphQLType> getResult() {
-            return result;
-        }
-    }
-
-
-
-    private static final TypeTraverser SHALLOW_TRAVERSER = new TypeTraverser((node) -> Collections.emptyList());
-
-    private static final TypeTraverser TRAVERSER = new TypeTraverser();
